requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
Django==5.2.6
django-cors-headers==4.3.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
dnspython==2.8.0
email-validator==2.3.0
filelock==3.20.0
fsspec==2025.12.0
gunicorn==21.2.0
idna==3.11
Jinja2==3.1.6
llvmlite==0.41.1
MarkupSafe==3.0.3
mpmath==1.3.0
mutagen==1.47.0
networkx==3.6
numpy==1.26.4
packaging==25.0
pillow==12.0.0
platformdirs==4.5.1
psycopg2-binary==2.9.9
pycparser==2.23
PyJWT==2.8.0
pymongo==4.6.1
python-decouple==3.8
pytz==2025.2
requests==2.32.5
soundfile==0.13.1
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.1.0
torchaudio==2.1.0
typing_extensions==4.15.0
urllib3==2.6.0
